requirements.txt
Flask
python-dotenv
botbuilder-core==4.14.2
botbuilder-schema==4.14.2
botbuilder-ai==4.14.2
botbuilder-dialogs==4.14.2
python-dateutil
price-parser
opencensus-ext-azure
aiohttp